split protocol.rs into directory module
